requirements.txt

Django>=4.0
google-cloud-bigquery>=3.0
requests>=2.25
celery>=5.2
redis>=4.0
django-celery-beat>=2.2
google-auth-oauthlib==1.0.0
google-auth==2.22.0
google-analytics-data==0.16.2
google-api-python-client==2.95.0
cryptography==41.0.3